requirements.txt
APScheduler==3.9.1
celery==5.2.7
Django==4.0
django-allauth==0.50.0
django-apscheduler==0.6.2
django-filter==21.1
flower==1.1.0
Pillow==9.1.0
python-dotenv==0.19.2
redis==4.3.4
